migration_app_to_ts: app, index
