+++ index.ts
@@ -1,5 +1,7 @@
-require('dotenv').config()
-const app = require('./src/app')
+import {config} from 'dotenv'
+import {app} from './src/app'
+
+config()
 
 const port = process.env.PORT
 app.listen(port, () => console.log(`Server is running on port ${port}...`))
