+++ package.json
@@ -21,6 +21,11 @@
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
     "pg": "^8.7.1",
+    "ts-node": "^10.2.1",
     "uuid": "^8.3.2"
+  },
+  "devDependencies": {
+    "tslint": "^6.1.3",
+    "typescript": "^4.4.3"
   }
 }
